requirements.txt
pandas
scikit-learn
matplotlib
numpy
torch
ucimlrepo
prequests 
websockets 
asyncio
pyarrow
python-dotenv
